samples: cellular: mmss: Add CoAP support
Rework MMSS to handle CoAP as well.

Reorganize how the message queue works so the raw data and appId
are queued, not the fully encoded message.  This uses less RAM
during periods of no connectivity, and makes CoAP and MQTT work
the same until the messages are dequeued and are ready to be sent.
For CoAP, the data itself is passed to service-specific APIs which
do the encoding for the caller, vs. MQTT where the sample must
first do the encoding, then transmit the encoded messages using
MQTT.

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>